Cleaned up creating custom entity docs
