+++ .changeset/short-lamps-wave.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend': patch
+---
+
+Refactored `doEmit` into a standalone utility function and exported it
